Move CreateConferenceFormTest test methods to ConferenceTest
